[TM-2208] refactor dashboard project connection to DashboardEntity and align structure with Entity pattern
