+++ src/connections/DashboardEntity.ts
